requirements.txt
selenium>=4.11.0
allure-pytest>=2.13.2
webdriver-manager>=4.0.0
